Can switch to big ship now
Apparently there is this thing called sleep that some people do.

Switching to big ship for testing now works, by changing the players
ship in update() instead of loading new level. Loading new levels
without restarting the whole game still needs to be sorted.

Added a maxVelocity variable to ships, and a little roundin' to align to
grid. Thing is, it's not just ships that need to be aligned, so gotta do
more inheriting inheriting the inheriting?

Added up/down movement for testing and stuff. Playing with it makes me
think that there will need to be an acceleration speed for everything so
that you can start slow then boost of the screen at end of level, or
missiles can start slow then speed up (so you can dodge them!)

Also check last commit!

The game will use space pounds as currency.