Press 3 to set your cockpit on fire
Or 4 to extinguish the flames. This is maybe way too flashy. need
opinions and performance tests. Flat grey or red looked kinda boring.

The 3 and 4 key events just delete themselves from the key array when
used, which is quite nice. Going to use this for menus where we don't
want the player tapping a button and having it pressed 9 times in game.
Still think cooldowns for weapons can be done by the weapons like: if
(keydown && playerShip.weapon.cooldown === 0).