Pixel based collision detection
Read the documentation.

Also you can now set all the sizing stuff from the 3 variables
gridSizeX, gridSizeY, cSize, in the JavaScript file. I would still
prefer 31x31, but showed it to a couple of people and they said keep
current size or bigger so it's certainly not going back to 31x31 and I
should shut up about it :P