Don't even know what I did to these images :P
Also I've been learning how to inheritance with JavaScript a lot over
the last 24 hours. I think if both of us come up with a way of managing
everything (the few hundred lines of spritey entity stuff), we can just
pick whichever implementation is best. I.e. I don't mind if I spend 8
hours doing it then yours is better so we go with yours :P More likely
that we'd come up with a mixture of the two though I think.

Won't be on TS again until Friday or Saturday, can use GitHub branches
and DropBox and Hangouts and that Drive Document if either of us come up
with something amazing.

Also Minecon tickets going on sale in 7 hours if you're interested. I'm
going to try and get a couple (but they're expensive and will go in 2
minutes).