Weapons should be dynamic
However the actual emitter does not follow the ship when it moves. This
would be made much easier if we could just pass in the position of the
ship once and it would change when the ships pos changes but that
doesn't. Going to try and update it inside the draw function (defo need
a more elegant way though)