Stuff now moves in intervals based on velocity
It looks like it works!

Basically there is a counter that counts each time an update is run,
every update objects with v = 1 move 1 px, every 2 updates those with v
= 0.5 move 1 px and every 4 updates those with v = 0.25 move 1 px.

This seems to make the entire screen update at the right time. It all is
hard coded right now so needs to be made dynamic (only objects with 1,
0.5 or 0.25 v move and only down)

I did change a for loop to go backwards so might try that in the master
branch.