We have a z-Buffer! (Woo!)
Entities now have a zIndex property. 0 is the layer that the player is
on, -1 is closer to the 'camera' and 1, 2 is into the background.

If you want to add another layer you will need to add it into the code
around line 784 onwards, the pattern should be easy enough to understand
;)